+++ app/src/main/java/com/peludosteam/ismarket/domain/Product.kt
@@ -3,7 +3,7 @@ package com.peludosteam.ismarket.domain
 data class Product(
     var id: String = "",
     var name: String = "",
-    var price: Double = 0.0,
+    var price: Int = 0,
     var description: String = "",
     val imageRes: String = "",
     var stock: Int = 0,
